@@ -20,26 +20,26 @@
     "@types/ejs": "^3.1.0",
     "axios": "^0.28.0",
     "blaise-api-node-client": "https://github.com/ONSdigital/blaise-api-node-client",
-    "blaise-design-system-react-components": "git+https://github.com/ONSdigital/blaise-design-system-react-components#0.13.2",
+    "blaise-design-system-react-components": "git+https://github.com/ONSdigital/blaise-design-system-react-components#0.14.1",
     "dotenv": "^16.4.1",
     "ejs": "^3.1.10",
     "express": "^4.20.0",
     "node-cache": "^5.1.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.2.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.0",
     "react-scripts": "^5.0.0",
-    "typescript": "^4.5.4"
+    "typescript": "~5.3.3"
   },
   "devDependencies": {
     "@babel/core": "^7.16.7",
     "@babel/preset-env": "^7.16.8",
     "@babel/preset-typescript": "^7.16.7",
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^12.1.2",
+    "@testing-library/react": "^14.2.1",
     "@types/jest": "^27.4.0",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^18.2.51",
+    "@types/react-dom": "^18.2.18",
     "@types/supertest": "^2.0.11",
     "axios-mock-adapter": "^1.20.0",
     "babel-jest": "^27.4.6",
