@@ -1,16 +1,22 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import App from "./app";
-//import * as serviceWorker from "./serviceWorker";
-import {BrowserRouter as Router} from "react-router-dom";
+import { BrowserRouter as Router } from "react-router-dom";
 
-ReactDOM.render(
+const container = document.getElementById("root");
+
+if (!container) {
+    throw new Error("Root container element not found");
+}
+
+const root = createRoot(container);
+
+root.render(
     <React.StrictMode>
         <Router>
-            <App/>
+            <App />
         </Router>
-    </React.StrictMode>,
-    document.getElementById("root")
+    </React.StrictMode>
 );
 
 // If you want your app to work offline and load faster, you can change
